Simple refactor to functional
